+++ src/context.rs
@@ -17,7 +17,6 @@ use timer;
 use GameError;
 use GameResult;
 
-
 /// A `Context` is an object that holds on to global resources.
 /// It basically tracks hardware state such as the screen, audio
 /// system, timers, and so on.  Generally this type is **not** thread-
@@ -152,7 +151,16 @@ impl Context {
         let sdl_context = sdl2::init()?;
         let mut fs = Filesystem::new(game_id, author)?;
 
-        let config = fs.read_config().unwrap_or(default_config);
+        let config = match fs.read_config() {
+            Ok(config) => {
+                info!("Loading conf.toml");
+                config
+            }
+            Err(e) => {
+                info!("Could not load conf.toml, using default: {:?}", e);
+                default_config
+            }
+        };
 
         Context::from_conf(config, fs, sdl_context)
     }
@@ -160,6 +168,7 @@ impl Context {
     /// Prints out information on the resources subsystem.
     pub fn print_resource_stats(&mut self) {
         // TODO: consider using logging macros.
+        // ACTUALLY TODO: Deprecate this in favor of filesystem.log_all()
         self.filesystem.print_all();
     }
 
@@ -282,10 +291,8 @@ impl ContextBuilder {
     }
 }
 
-
-
 #[cfg(debug_assertions)]
-use std::sync::atomic::{ATOMIC_USIZE_INIT, AtomicUsize, Ordering};
+use std::sync::atomic::{AtomicUsize, Ordering, ATOMIC_USIZE_INIT};
 #[cfg(debug_assertions)]
 static DEBUG_ID_COUNTER: AtomicUsize = ATOMIC_USIZE_INIT;
 
@@ -293,9 +300,8 @@ static DEBUG_ID_COUNTER: AtomicUsize = ATOMIC_USIZE_INIT;
 /// is contained in each thing created from the Context which contains
 /// data that becomes invalid when the Context goes away (ie, texture
 /// handles).  When compiling without assertions (ie in release mode) it
-/// is replaced with a zero-size type, compiles down to nothing, and
-/// since all comparisons with it will be true, should disappear entirely
-/// with a puff of logic.
+/// is replaced with a zero-size type, compiles down to nothing,
+/// and should disappear entirely with a puff of optimization logic.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 #[cfg(debug_assertions)]
 pub(crate) struct DebugId(u32);
@@ -312,7 +318,7 @@ impl DebugId {
         assert!(DEBUG_ID_COUNTER.load(Ordering::SeqCst) as u32 > id);
         DebugId(id)
     }
-    
+
     pub fn get(ctx: &Context) -> Self {
         DebugId(ctx.debug_id.0)
     }
@@ -322,7 +328,6 @@ impl DebugId {
             panic!("Tried to use a resource with a Context that did not create it; this should never happen!");
         }
     }
-            
 }
 
 #[cfg(not(debug_assertions))]
@@ -331,7 +336,6 @@ impl DebugId {
         DebugId
     }
 
-    
     pub fn get(_ctx: &Context) -> Self {
         DebugId
     }
@@ -340,4 +344,3 @@ impl DebugId {
         // Do nothing.
     }
 }
-