+++ src/error.rs
@@ -259,6 +259,3 @@ impl From<gfx::shade::ProgramError> for GameError {
         GameError::ShaderProgramError(e)
     }
 }
-
-
-