+++ src/graphics/image.rs
@@ -7,7 +7,7 @@ use image;
 
 use graphics::*;
 use graphics::shader::*;
-use context::{DebugId, Context};
+use context::{Context, DebugId};
 use GameResult;
 use GameError;
 
@@ -174,7 +174,13 @@ impl Image {
         // TODO: Check for overflow on 32-bit systems here
         let expected_bytes = width as usize * height as usize * 4;
         if expected_bytes != rgba.len() {
-            let msg = format!("Tried to create a texture of size {}x{}, but gave {} bytes of data (expected {})", width, height, rgba.len(), expected_bytes);
+            let msg = format!(
+                "Tried to create a texture of size {}x{}, but gave {} bytes of data (expected {})",
+                width,
+                height,
+                rgba.len(),
+                expected_bytes
+            );
             return Err(GameError::ResourceLoadError(msg));
         }
         let kind = gfx::texture::Kind::D2(width, height, gfx::texture::AaMode::Single);
@@ -190,7 +196,7 @@ impl Image {
             blend_mode: None,
             width: u32::from(width),
             height: u32::from(height),
-            debug_id
+            debug_id,
         })
     }
 
@@ -315,6 +321,6 @@ mod tests {
         let ctx = &mut Context::load_from_conf("unittest", "unittest", c).unwrap();
         let _i = assert!(Image::from_rgba8(ctx, 0, 0, &vec![]).is_err());
         let _i = assert!(Image::from_rgba8(ctx, 3432, 432, &vec![]).is_err());
-        let _i = Image::from_rgba8(ctx, 2, 2, &vec![99;16]).unwrap();
+        let _i = Image::from_rgba8(ctx, 2, 2, &vec![99; 16]).unwrap();
     }
 }