+++ src/graphics/spritebatch.rs
@@ -60,7 +60,7 @@ impl SpriteBatch {
 
     /// Adds a new sprite to the sprite batch.
     ///
-    /// Returns a handle with whictypeh to modify the sprite using `set()`
+    /// Returns a handle with which type to modify the sprite using `set()`
     pub fn add(&mut self, param: graphics::DrawParam) -> SpriteIdx {
         self.sprites.push(param);
         SpriteIdx(self.sprites.len() - 1)
@@ -158,10 +158,7 @@ impl SpriteBatch {
     #[deprecated]
     #[allow(deprecated)]
     pub fn with_image<'a>(&'a mut self, image: &'a graphics::Image) -> BoundSpriteBatch<'a> {
-        BoundSpriteBatch {
-            image,
-            batch: self,
-        }
+        BoundSpriteBatch { image, batch: self }
     }
 }